==> code/01_DataDownload.Rmd <==
---
title: "Download and Merge RECS Data"
author: "Arfa Aijazi"
date: "October 2023"
output:
  html_document:
    df_print: paged
---

This script downloads and merges the RECS 2015 and RECS 2020 data sets 

### Setup
Load libraries
```{r echo=TRUE, message=FALSE, warning=FALSE}
library(tidyverse)
library(knitr)
library(scales)

options(es.use_symbols = TRUE)
```

### Data processing
Load raw microdata files from the Energy Information Administration (EIA) website
```{r echo=TRUE, message=FALSE, warning=FALSE}
# Pull data from EIA website
# data_2015 = "https://www.eia.gov/consumption/residential/data/2015/csv/recs2015_public_v4.csv"
# data_2020 = "https://www.eia.gov/consumption/residential/data/2020/csv/recs2020_public_v5.csv"

# Predownloaded data file
data_2015 = "../data/recs2015_public_v4.csv"
data_2020 = "../data/recs2020_public_v5.csv"

recs_2015 <- read_csv(data_2015)
recs_2020 <- read_csv(data_2020)
```

Subset variables of interest and prepare data for merging
```{r echo=TRUE, message=FALSE, warning=FALSE}
# Prepare 2015 data set
recs_2015_sub <- recs_2015 %>%
  select(CDD30YR, CDD65, HDD30YR, HDD65, DBT1, DBT99, HOUSEHOLDER_RACE, SDESCENT, HHAGE, NUMADULT, NUMCHILD, MONEYPY, EMPLOYHH, EDUCATION, KOWNRENT, ELPAY, NGPAY, LPGPAY, FOPAY, UATYP10, TYPEHUQ, YEARMADERANGE, WALLTYPE, ROOFTYPE, ADQINSUL, DRAFTY, TOTROOMS, WINDOWS, WINFRAME, TYPEGLASS, AIRCOND, HEATHOME, COOLTYPE, SWAMPCOL, EQUIPM,  NOACDAYS, NOHEATDAYS, AGECENAC, WWACAGE, EQUIPAGE, NOACBROKE, NOHEATBROKE, NOACEL, NOHEATEL, NOHEATNG, NOHEATBULK, NUMCFAN, NUMFLOORFAN, BACKUP, SOLAR, SCALEG, SCALEB, SCALEE, HOTMA, COLDMA, NWEIGHT) %>%
  mutate(NUMADULT2 = 0) %>%
  mutate(MONEYPY = case_when(MONEYPY == 1 ~ 0,
                             MONEYPY == 2 ~ 20000,
                             MONEYPY == 3 ~ 40000,
                             MONEYPY == 4 ~ 60000,
                             MONEYPY == 5 ~ 80000,
                             MONEYPY == 6 ~ 100000,
                             MONEYPY == 7 ~ 120000,
                             MONEYPY == 8 ~ 140000)) %>%
  mutate(WALLTYPE = ifelse(WALLTYPE == 9, 99, WALLTYPE)) %>%
  mutate(ROOFTYPE = ifelse(ROOFTYPE == 7, 6, ROOFTYPE)) %>% # recode concrete tile roof
  mutate(ROOFTYPE = ifelse(ROOFTYPE == 9, 99, ROOFTYPE)) %>%
  mutate(ADQINSUL = ADQINSUL*-1+4) %>% # recode to increasing level of insulation
  mutate(WINDOWS = case_when(WINDOWS == 10 ~ 1,
                             WINDOWS == 20 ~ 2,
                             WINDOWS == 30 ~ 3,
                             WINDOWS == 41 ~ 4,
                             WINDOWS == 42 ~ 5,
                             WINDOWS == 50 ~ 6,
                             WINDOWS == 60 ~ 7)) %>%
  mutate(COOLTYPE = case_when(SWAMPCOL == 1 & COOLTYPE > 0 ~ 3, # Swamp cooler used in addition to another system
                               SWAMPCOL == 1 & COOLTYPE < 0 ~ 6, # Primarily uses swamp cooler
                               TRUE ~ COOLTYPE)) %>% # Else keeps original value
  mutate(EQUIPM = ifelse(EQUIPM == 21, 99, EQUIPM)) %>% # Sets value to 99 if the heating system type is other
  mutate(AGECENAC = case_when(AGECENAC == 5 ~ 6, # recode > 20 years to 6
                              AGECENAC == 42 ~ 5, # recode 15-19 years to 5
                              AGECENAC == 41 ~ 4, # recode 10-14 years to 4
                              TRUE ~ AGECENAC)) %>% 
  mutate(WWACAGE = case_when(WWACAGE == 5 ~ 6, # recode > 20 years to 6
                              WWACAGE == 42 ~ 5, # recode 15-19 years to 5
                              WWACAGE == 41 ~ 4, # recode 10-14 years to 4
                              TRUE ~ WWACAGE)) %>% 
  mutate(ACEQUIPAGE = pmax(AGECENAC, WWACAGE)) %>% # reports the age of the older air conditioning system
  mutate(EQUIPAGE = case_when(EQUIPAGE == 5 ~ 6, # recode > 20 years to 6
                              EQUIPAGE == 42 ~ 5, # recode 15-19 years to 5
                              EQUIPAGE == 41 ~ 4, # recode 10-14 years to 4
                              TRUE ~ EQUIPAGE)) %>%
  mutate(POVERTY = case_when(MONEYPY <= 20000 ~ 1,
                             NUMCHILD + NUMCHILD >= 4 & MONEYPY <= 40000 ~ 1, 
                             TRUE ~ 0)) %>% # based on poverty thresholds for 2015 by family size
  select(-AGECENAC, -WWACAGE, - SWAMPCOL) %>%
  mutate(YEAR = 2015)
  
# Prepare 2020 data set
recs_2020_sub <- recs_2020 %>%
    select(CDD30YR_PUB, CDD65, HDD30YR_PUB, HDD65, DBT1, DBT99, HOUSEHOLDER_RACE, SDESCENT, HHAGE, NUMADULT1, NUMADULT2, NUMCHILD, MONEYPY, EMPLOYHH, EDUCATION, KOWNRENT, ELPAY, NGPAY, LPGPAY, FOPAY, UATYP10, TYPEHUQ, YEARMADERANGE, WALLTYPE, ROOFTYPE, ADQINSUL, DRAFTY, TOTROOMS, WINDOWS, WINFRAME, TYPEGLASS, AIRCOND, HEATHOME, ACEQUIPM_PUB, ACEQUIPAUXTYPE_PUB, EQUIPM, NOACDAYS, NOHEATDAYS, ACEQUIPAGE, EQUIPAGE, NOACBROKE, NOHEATBROKE, NOACEL, NOHEATEL, NOHEATNG, NOHEATBULK, NUMCFAN, NUMFLOORFAN, BACKUP, SOLAR, SCALEG, HOTMA, COLDMA, NWEIGHT, TREESHAD) %>%
  mutate(NUMADULT = NUMADULT1 + NUMADULT2) %>%
  rename(CDD30YR = CDD30YR_PUB, HDD30YR = HDD30YR_PUB) %>%
  mutate(HOUSEHOLDER_RACE = ifelse(HOUSEHOLDER_RACE == 6, 7, HOUSEHOLDER_RACE)) %>%
  mutate(MONEYPY = case_when(MONEYPY == 1 ~ 0,
                             MONEYPY == 2 ~ 5000,
                             MONEYPY == 3 ~ 75000,
                             MONEYPY == 4 ~ 10000,
                             MONEYPY == 5 ~ 12500,
                             MONEYPY == 6 ~ 15000,
                             MONEYPY == 7 ~ 20000,
                             MONEYPY == 8 ~ 25000,
                             MONEYPY == 9 ~ 30000,
                             MONEYPY == 10 ~ 35000,
                             MONEYPY == 11 ~ 40000,
                             MONEYPY == 12 ~ 50000,
                             MONEYPY == 13 ~ 60000,
                             MONEYPY == 14 ~ 75000,
                             MONEYPY == 15 ~ 100000,
                             MONEYPY == 16 ~ 150000)) %>%
  mutate(EMPLOYHH = case_when(EMPLOYHH > 2 ~ 0,
                              TRUE ~ EMPLOYHH)) %>%
  mutate(ADQINSUL = ADQINSUL*-1+4) %>% # recode to increasing level of insulation
  mutate(COOLTYPE = case_when(ACEQUIPM_PUB == 1 ~ 1, # Primarily uses central air conditioner
                              ACEQUIPM_PUB >= 3 & ACEQUIPM_PUB <= 5 ~ 2, # Primarily uses ductless heat pumps, window or wall conditioner, and portable air conditioner
                              ACEQUIPM_PUB == 6 ~ 6, # Primarily uses evaporative or swamp cooler
                              ACEQUIPM_PUB > 0 & ACEQUIPAUXTYPE_PUB > 0 ~ 3,
                              TRUE ~ -2)) %>%
  mutate(POVERTY = case_when(MONEYPY <= 10000 ~ 1,
                             NUMADULT == 1 & HHAGE >= 65 & MONEYPY <= 12500 ~ 1,
                             NUMADULT == 2 & HHAGE >= 65 & MONEYPY <= 15000 ~ 1,
                             NUMCHILD == 1 & MONEYPY <= 20000 ~ 1,
                             NUMADULT + NUMCHILD <= 3 & MONEYPY <= 25000 ~ 1,
                             NUMADULT + NUMCHILD <= 4 & MONEYPY <= 30000 ~ 1,
                             NUMADULT + NUMCHILD <= 5 & MONEYPY <= 35000 ~ 1,
                             NUMADULT + NUMCHILD <= 6 & MONEYPY <= 40000 ~ 1,
                             NUMADULT + NUMCHILD <= 8 & MONEYPY <= 50000 ~ 1,
                             NUMADULT + NUMCHILD <= 9 & MONEYPY <= 60000 ~ 1, 
                             NUMADULT + NUMCHILD == 6 & NUMCHILD >= 4 & MONEYPY <= 35000 ~ 1,
                             TRUE ~ 0)) %>% # based on poverty thresholds for 2020 by family size
  select(-NUMADULT1, -ACEQUIPM_PUB, -ACEQUIPAUXTYPE_PUB) %>%
  mutate(YEAR = 2020) 
```

Merge 2015 and 2020 data sets and group temperature-related morbidity
```{r}
recs_all <- full_join(recs_2015_sub, recs_2020_sub) %>%
  mutate(TEMPMA = ifelse(HOTMA == 1 | COLDMA == 1, 1, 0))
```

```{r}
recs_recode <- recs_all %>%
  mutate(WHITE = ifelse(HOUSEHOLDER_RACE == 1, 1, 0)) %>%
  mutate(BLACK = ifelse(HOUSEHOLDER_RACE == 2, 1, 0)) %>%
  mutate(ASIAN = ifelse(HOUSEHOLDER_RACE == 4, 1, 0)) %>%
  mutate(MIXED = ifelse(HOUSEHOLDER_RACE == 7, 1, 0)) %>%
  mutate(OTHER_RACE = ifelse(HOUSEHOLDER_RACE == 3 | HOUSEHOLDER_RACE == 5 | HOUSEHOLDER_RACE == 6, 1, 0)) %>%
  mutate(OVER65 = ifelse(HHAGE > 65 | NUMADULT2 > 0, 1, 0)) %>% # ADULT2 in the 2020 data set counts the number of adults over 65
  mutate(NHSLDMEM = NUMADULT + NUMCHILD) %>%
  mutate(LIVESALONE = ifelse(NHSLDMEM == 1, 1, 0)) %>%
  mutate(LARGEHOUSE = ifelse(NHSLDMEM >= 7, 1, 0) ) %>%
  mutate(HIGHSCHOOL = ifelse(EDUCATION <= 2, 1, 0)) %>%
  mutate(RENT = ifelse(KOWNRENT == 2, 1, 0)) %>%
  mutate(ELPAY = ifelse(ELPAY == 1, 1, 0)) %>%
  mutate(NGPAY = ifelse(NGPAY == 1, 1, 0)) %>%
  mutate(FOPAY = ifelse(FOPAY == 1, 1, 0)) %>%
  mutate(LPGPAY = ifelse(LPGPAY == 1, 1, 0)) %>%
  mutate(PAYSUTL = ifelse(ELPAY == 1 | NGPAY == 1 | FOPAY == 1 | LPGPAY, 1, 0)) %>%
  mutate(APARTMENT = ifelse(TYPEHUQ >= 4, 1, 0)) %>%
  mutate(YEARMADERANGE = case_when(YEARMADERANGE == 1 ~ 1950,
                                   YEARMADERANGE == 2 ~ 1959,
                                   YEARMADERANGE == 3 ~ 1969,
                                   YEARMADERANGE == 4 ~ 1979,
                                   YEARMADERANGE == 5 ~ 1989,
                                   YEARMADERANGE == 6 ~ 1999,
                                   YEARMADERANGE == 7 ~ 2009,
                                   YEARMADERANGE == 8 ~ 2015,
                                   YEARMADERANGE == 9 ~ 2020)) %>%
  mutate(WINDOWS = case_when(WINDOWS == 1 ~ 1,
                             WINDOWS == 2 ~ 3,
                             WINDOWS == 3 ~ 6,
                             WINDOWS == 4 ~ 10,
                             WINDOWS == 5 ~ 16,
                             WINDOWS == 6 ~ 20,
                             WINDOWS == 7 ~ 30)) %>%
  mutate(MOBILE = ifelse(TYPEHUQ == 1, 1, 0)) %>%
  mutate(WALLTYPE = case_when(WALLTYPE == 1 ~ "Brick",
                              WALLTYPE == 2 ~ "Wood",
                              WALLTYPE == 3 ~ "Siding",
                              WALLTYPE == 4 ~ "Stucco",
                              WALLTYPE == 5 ~ "Shingle",
                              WALLTYPE == 6 ~ "Stone",
                              WALLTYPE == 7 ~ "Concrete",
                              WALLTYPE == 99 ~ "Other")) %>%
  mutate(THERMALMASS.WALL = case_when(WALLTYPE == "Brick" ~ 0.143, # Based on thermal mass calculation in separate spreadsheet
                                      WALLTYPE == "Wood" ~ 0.013,
                                      WALLTYPE == "Siding" ~ 0.004,
                                      WALLTYPE == "Stucco" ~ 0.024,
                                      WALLTYPE == "Stone" ~ 0.202,
                                      WALLTYPE == "Concrete" ~ 0.410,
                                      TRUE ~ 0)) %>%
  mutate(ROOFTYPE = case_when(ROOFTYPE == 1 ~ "CeramicTiles",
                              ROOFTYPE == 2 ~ "WoodShingle",
                              ROOFTYPE == 3 ~ "Metal",
                              ROOFTYPE == 4 ~ "SlateShake",
                              ROOFTYPE == 5 ~ "Shingles",
                              ROOFTYPE == 6 ~ "ConcreteTiles",
                              ROOFTYPE == 99 ~ "Other",
                              ROOFTYPE == -2 ~ "NoRoof")) %>%
  mutate(THERMALMASS.ROOF = case_when(ROOFTYPE == "CeramicTiles" ~ 0.008, # Based on thermal mass calculation in separate spreadsheet
                                      ROOFTYPE == "WoodShingle" ~ 0.013,
                                      ROOFTYPE == "Metal" ~ 0.001,
                                      ROOFTYPE == "SlateShake" ~ 0.044,
                                      ROOFTYPE == "Shingles" ~ 0.069,
                                      ROOFTYPE == "ConcreteTiles" ~ 0.038,
                                      TRUE ~ 0)) %>%
  mutate(THERMALMASS.WALL = ifelse(ADQINSUL <= 2, THERMALMASS.WALL, 0)) %>%
  mutate(THERMALMASS.ROOF = ifelse(ADQINSUL <= 2, THERMALMASS.ROOF, 0)) %>%
  mutate(COOLTYPE = case_when(COOLTYPE == 1 ~ 1,
                              COOLTYPE == 2 ~ 0.5,
                              COOLTYPE == 3 ~ 1,
                              COOLTYPE == 6 ~ 0.5,
                              COOLTYPE == -2 & AIRCOND == 0 ~ 0)) %>%
  mutate(EQUIPM = case_when(EQUIPM == 2 ~ 1, # Steam or hot water system with radiators or pipes
                            EQUIPM == 3 ~ 1, # Central furnance
                            EQUIPM == 4 ~ 1, # Central heat pump
                            EQUIPM == 5 ~ 1, # Built-in electric units, installed in walls, ceilings, baseboards, or floors
                            EQUIPM == 6 ~ 1, # Built-in floor/wall pipeless furnace
                            EQUIPM == 7 ~ 1, # Built-in room heater burning gas, oil, or kerosene
                            EQUIPM == 8 ~ 0.5, # Wood-burning stove
                            EQUIPM == 9 ~ 0.5, # Fireplace
                            EQUIPM == 10 ~ 0.5, # Portable electric heater
                            EQUIPM == 13 ~ 1, # Ductless heat pump
                            EQUIPM == 99 ~ 0, 
                            EQUIPM == -2 ~ 0)) %>%
  mutate(EMPLOYHH = ifelse(EMPLOYHH == 2, 0.5, EMPLOYHH)) %>% # Part-time employment
  mutate(UNEMPLOYED = -1*EMPLOYHH) %>%
  mutate(DRAFTY = (DRAFTY*-1) + 4) %>%
  mutate(OFFGRID = ifelse(BACKUP == 1 | SOLAR == 1, 1, 0)) %>%
  mutate(HVAC_OPERATION = case_when(NOACEL == 1 ~ 1,
                               NOHEATEL == 1 ~ 1,
                               NOHEATNG == 1 ~ 1,
                               NOHEATBULK == 1 ~ 1,
                               SCALEB != 0 ~ 1, 
                               SCALEG != 0 ~ 1,
                               SCALEE != 0 ~ 1,
                               TRUE ~ 0
                               )) %>%
  mutate(HVAC_MAINTENANCE = ifelse(NOACBROKE == 1 | NOHEATBROKE == 1, 1, 0)) %>%
  mutate(WWR = WINDOWS/TOTROOMS) %>%
  mutate(NUMFANS = NUMCFAN + NUMFLOORFAN)
```

```{r}
recs_subset <- recs_recode %>%
  select(DBT1, DBT99, WHITE, BLACK, ASIAN, MIXED, OTHER_RACE, SDESCENT, OVER65, LIVESALONE, LARGEHOUSE, POVERTY, UNEMPLOYED, HIGHSCHOOL, RENT, ELPAY, NGPAY, LPGPAY, FOPAY, APARTMENT, MOBILE, YEARMADERANGE, THERMALMASS.WALL, THERMALMASS.ROOF, ADQINSUL, DRAFTY, WWR, TYPEGLASS, COOLTYPE, EQUIPM, HVAC_OPERATION, HVAC_MAINTENANCE, NUMFANS, OFFGRID, TEMPMA)
```

### Export
```{r}
write_csv(recs_subset, "../data/recs_subset.csv")
```
