==> code/03_DataProcessing.Rmd <==
---
title: "Data Processing"
author: "Arfa Aijazi"
date: "November 2023"
output: html_document
---

This script prepares the merged RECS data in preparation for machine learning

### Setup
Load libraries
```{r echo=TRUE, message=FALSE, warning=FALSE}
library(tidyverse)
library(caret)
library(rstatix)

options(es.use_symbols = TRUE)
```

Custom plot theme
```{r message=FALSE, warning=FALSE, include=FALSE}
theme_custom = function() {
  theme_minimal() %+replace%
    theme(legend.position = "top") +
    theme(panel.grid = element_blank()) +
    theme(strip.background = element_blank()) +
    theme(panel.border = element_blank()) +
    theme(panel.background = element_blank()) +
    # theme(legend.title = element_blank()) +
    theme(axis.title = element_blank()) +
    theme(text = element_text(size = 7, colour = "#000000"))
}
```

Import data
```{r}
recs_subset <- read.csv("../data/recs_subset.csv", stringsAsFactors = TRUE)

y <- recs_subset$TEMPMA
```

Zero- and Near Zero-Variance Predictors

```{r}
nzv <- nearZeroVar(recs_subset, saveMetrics = TRUE, freqCut = 95, uniqueCut = 0.008270615) %>% # Look for variables with less variance than TEMPMA
  rownames_to_column(var = "Variable")
write_csv(nzv, "../data/recs_nzv.csv")
# nzv <- nearZeroVar(recs_subset)
# recs_nzv <- recs_subset[,-nzv]
# recs_nzv$TEMPMA <- y # add TEMPMA back
```
None of the variables met the threshold for removal


Correlated predictors
```{r}
descrCor <-  cor(recs_subset, method = "spearman")

df_Cor <- as.data.frame(descrCor) %>%
  rownames_to_column(var = "Variable")

write_csv(df_Cor, "../data/recs_cor.csv")

summary(descrCor[upper.tri(descrCor)])

# descrCor <-  cor(recs_nzv, method = "pearson")
# summary(descrCor[upper.tri(descrCor)])

highlyCorDescr <- findCorrelation(descrCor, cutoff = 0.75)
```
None of the variables met the threshold for removal

Plot correlation matrix
```{r}
corr <- cor_mat(recs_subset, method = "spearman") %>%
  pull_upper_triangle(diagonal = TRUE)

write_csv(corr, "../data/recs_cor.csv")
```


Linear dependencies
```{r}
comboInfo <- findLinearCombos(recs_subset)
```

No linear combinations

Centering and scaling
```{r}
recs_input <- recs_subset

preProcValues <- preProcess(recs_input, method = c("range"))

recs_standardized <- predict(preProcValues, recs_input)
```

### Export
```{r}
write_csv(recs_standardized, "../data/recs_standardized.csv")
```
